# PJM/contracts/openapi/projectmind-api.v1.json
{
  "components": {
    "schemas": {
...
      "CreateIntegrationRequest": {
        "additionalProperties": false,
        "description": "登録済み Provider instance を作成する request。",
        "properties": {
          "capabilities": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Capabilities",
            "type": "array"
          },
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          },
          "kind": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "provider": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Provider",
            "type": "string"
          },
          "scope": {
            "additionalProperties": true,
            "title": "Scope",
            "type": "object"
          },
          "secret_reference_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secret Reference Id"
          }
        },
        "required": [
          "name",
          "kind",
          "provider",
          "capabilities",
          "scope",
          "config"
        ],
        "title": "CreateIntegrationRequest",
        "type": "object"
      },
...
      "DisableIntegrationRequest": {
        "additionalProperties": false,
        "description": "Optimistic revision 付き Integration disable request。",
        "properties": {
          "expected_revision": {
            "minimum": 1.0,
            "title": "Expected Revision",
            "type": "integer"
          }
        },
        "required": [
          "expected_revision"
        ],
        "title": "DisableIntegrationRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IntegrationListResponse": {
        "description": "Project の Integration 一覧。",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IntegrationResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "IntegrationListResponse",
        "type": "object"
      },
      "IntegrationResponse": {
        "description": "接続設定値と Secret locator を除外した Integration response。",
        "properties": {
          "capabilities": {
            "items": {
              "type": "string"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "config_keys": {
            "items": {
              "type": "string"
            },
            "title": "Config Keys",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "format": "uuid",
            "title": "Created By",
            "type": "string"
          },
          "disabled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disabled At"
          },
          "integration_id": {
            "format": "uuid",
            "title": "Integration Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "scope": {
            "additionalProperties": true,
            "title": "Scope",
            "type": "object"
          },
          "secret_reference_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secret Reference Id"
          },
          "status": {
            "$ref": "#/components/schemas/IntegrationStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "integration_id",
          "project_id",
          "name",
          "kind",
          "provider",
          "status",
          "revision",
          "capabilities",
          "scope",
          "config_keys",
          "secret_reference_id",
          "created_by",
          "created_at",
          "updated_at",
          "disabled_at"
        ],
        "title": "IntegrationResponse",
        "type": "object"
      },
...
      "ResourceBindingListResponse": {
        "description": "Project 内 ResourceBinding 一覧。",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ResourceBindingResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ResourceBindingListResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ProjectMind API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project_id}/integrations": {
      "get": {
        "description": "ADMIN に Project の Integration metadata を返す。",
        "operationId": "list_integrations_api_v1_projects__project_id__integrations_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Integrations",
        "tags": [
          "integrations"
        ]
      },
      "post": {
        "description": "ADMIN が登録済み Provider の Project instance を作成する。",
        "operationId": "create_integration_api_v1_projects__project_id__integrations_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "title": "X-Csrf-Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIntegrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Integration",
        "tags": [
          "integrations"
        ]
      }
    },
    "/api/v1/projects/{project_id}/integrations/{integration_id}/disable": {
      "post": {
        "description": "ADMIN が optimistic revision を指定して Integration を無効化する。",
        "operationId": "disable_integration_api_v1_projects__project_id__integrations__integration_id__disable_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "integration_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Integration Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "title": "X-Csrf-Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisableIntegrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Disable Integration",
        "tags": [
          "integrations"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/resource-bindings": {
      "get": {
        "description": "ADMIN に Project の binding と immutable Run snapshot を返す。",
        "operationId": "list_resource_bindings_api_v1_projects__project_id__resource_bindings_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceBindingListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Resource Bindings",
        "tags": [
          "integrations"
        ]
      },
      "put": {
        "description": "ADMIN が Project default または Task override binding を保存する。",
        "operationId": "put_resource_binding_api_v1_projects__project_id__resource_bindings_put",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "title": "X-Csrf-Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutResourceBindingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceBindingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Resource Binding",
        "tags": [
          "integrations"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/secret-references": {
      "get": {
        "description": "ADMIN に locator を含まない SecretReference metadata を返す。",
        "operationId": "list_secret_references_api_v1_projects__project_id__secret_references_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecretReferenceListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Secret References",
        "tags": [
          "integrations"
        ]
      },
      "post": {
        "description": "ADMIN が locator(ENVIRONMENT/FILE)または明文(MANAGED)で SecretReference を登録する。\n\nMANAGED の明文はサーバが即座に KEK 封入し、平文のまま保存・返却・ログ化しない。",
        "operationId": "create_secret_reference_api_v1_projects__project_id__secret_references_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "title": "X-Csrf-Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSecretReferenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecretReferenceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Secret Reference",
        "tags": [
          "integrations"
        ]
      }
    },
    "/api/v1/projects/{project_id}/secret-references/{secret_reference_id}/disable": {
      "post": {
        "description": "ADMIN が SecretReference を物理削除せず無効化する。",
        "operationId": "disable_secret_reference_api_v1_projects__project_id__secret_references__secret_reference_id__disable_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "secret_reference_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Secret Reference Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "title": "X-Csrf-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecretReferenceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Disable Secret Reference",
        "tags": [
          "integrations"
        ]
      }
    },
...
  }
}